src/apps/ocioview/requirements.txt
imageio
numpy
pygments
PyOpenGL
PySide6
QtAwesome
OpenColorIO